Clear some commentaries, move pattern-classes into separate folder.
