+++ CarRental.BussinessLayer/Managers/ServiceManager.cs
@@ -8,6 +8,7 @@ using CarRental.Data.Enums;
 using CarRental.Data.Interfaces;
 using CarRental.Data.Models.Car;
 using CarRental.Data.Models.Car.RecordTypes;
+using CarRental.Data.Models.Car.Seeds;
 
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
