+++ CarRental.BussinessLayer/Managers/SupplementDataInitializator.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using CarRental.Data.Enums;
 using CarRental.Data.Models.Car;
 using CarRental.Data.Models.Car.RecordTypes;
+using CarRental.Data.Models.Car.Seeds;
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
 // HOMEWORK: "ДЗ 4. Extension methods & Record type."
