+++ CarRental.PresentationLayer/Interfaces/ICarSelectivePattern.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using CarRental.Data.Models.Car;
+using CarRental.Data.Models.Car.Seeds;
 
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
