+++ CarRental.PresentationLayer/Models/Car/Depot.cs
@@ -7,6 +7,7 @@ using System.Drawing;
 using System.Diagnostics.CodeAnalysis;
 using CarRental.Data.Enums;
 using CarRental.Data.Interfaces;
+using CarRental.Data.Models.Car.Seeds;
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
 // HOMEWORK: "ДЗ 4. Extension methods & Record type."
