+++ CarRental.PresentationLayer/Models/Car/PatternInitializator.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using CarRental.Data.Models.Car.RecordTypes;
+using CarRental.Data.Models.Car.Seeds;
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
 // HOMEWORK: "ДЗ 4. Extension methods & Record type."
@@ -190,72 +191,6 @@ public class PatternInitializator
                 PitchInitial = dataWarehouse.ModelPatternDataDictionary[model][50],
                 PitchEnd = dataWarehouse.ModelPatternDataDictionary[model][51]
             }
-
-            //StatusInitialIndex = dataWarehouse.ModelPatternDataDictionary[model][0],
-            //StatusEndIndex = dataWarehouse.ModelPatternDataDictionary[model][1],
-
-            //EngineStatusInitialIndex = dataWarehouse.ModelPatternDataDictionary[model][2],
-            //EngineStatusEndIndex = dataWarehouse.ModelPatternDataDictionary[model][3],
-
-            //TransmissionStatusInitialIndex = dataWarehouse.ModelPatternDataDictionary[model][4],
-            //TransmissionStatusEndIndex = dataWarehouse.ModelPatternDataDictionary[model][5],
-
-            //InteriorStatusInitialIndex = dataWarehouse.ModelPatternDataDictionary[model][6],
-            //InteriorStatusEndIndex = dataWarehouse.ModelPatternDataDictionary[model][7],
-
-            //WheelsStatusInitialIndex = dataWarehouse.ModelPatternDataDictionary[model][8],
-            //WheelsStatusEndIndex = dataWarehouse.ModelPatternDataDictionary[model][9],
-
-            //LightsStatusInitialIndex = dataWarehouse.ModelPatternDataDictionary[model][10],
-            //LightsStatusEndIndex = dataWarehouse.ModelPatternDataDictionary[model][11],
-
-            //SignalStatusInitialIndex = dataWarehouse.ModelPatternDataDictionary[model][12],
-            //SignalStatusEndIndex = dataWarehouse.ModelPatternDataDictionary[model][13],
-
-            //ColorCarInitial = dataWarehouse.ModelPatternDataDictionary[model][14],
-            //ColorCarEnd = dataWarehouse.ModelPatternDataDictionary[model][15],
-            //PriceCarInitial = dataWarehouse.ModelPatternDataDictionary[model][16],
-            //PriceCarEnd = dataWarehouse.ModelPatternDataDictionary[model][17],
-            //MaxFuelCapacityInitial = dataWarehouse.ModelPatternDataDictionary[model][18],
-            //MaxFuelCapacityEnd = dataWarehouse.ModelPatternDataDictionary[model][19],
-            //CurrentFuelInitial = dataWarehouse.ModelPatternDataDictionary[model][20],
-            //CurrentFuelEnd = dataWarehouse.ModelPatternDataDictionary[model][21],
-            //SpeedCoeficientInitial = dataWarehouse.ModelPatternDataDictionary[model][22],
-            //SpeedCoeficientEnd = dataWarehouse.ModelPatternDataDictionary[model][23],
-
-            //AverageFuelConsumptionInitial = dataWarehouse.ModelPatternDataDictionary[model][24],
-            //AverageFuelConsumptionEnd = dataWarehouse.ModelPatternDataDictionary[model][25],
-            //FuelInitial = dataWarehouse.ModelPatternDataDictionary[model][26],
-            //FuelEnd = dataWarehouse.ModelPatternDataDictionary[model][27],
-            //TypeEngineInitial = dataWarehouse.ModelPatternDataDictionary[model][28],
-            //TypeEngineEnd = dataWarehouse.ModelPatternDataDictionary[model][29],
-            //PowerEngineInitial = dataWarehouse.ModelPatternDataDictionary[model][30],
-            //PowerEngineEnd = dataWarehouse.ModelPatternDataDictionary[model][31],
-
-            //TypeTransmissionInitial = dataWarehouse.ModelPatternDataDictionary[model][32],
-            //TypeTransmissionEnd = dataWarehouse.ModelPatternDataDictionary[model][33],
-            //SpeedCountInitial = dataWarehouse.ModelPatternDataDictionary[model][34],
-            //SpeedCountEnd = dataWarehouse.ModelPatternDataDictionary[model][35],
-
-            //ColorInteriorInitial = dataWarehouse.ModelPatternDataDictionary[model][36],
-            //ColorInteriorEnd = dataWarehouse.ModelPatternDataDictionary[model][37],
-            //MaterialInteriorInitial = dataWarehouse.ModelPatternDataDictionary[model][38],
-            //MaterialInteriorEnd = dataWarehouse.ModelPatternDataDictionary[model][39],
-
-            //MaterialWheelsInitial = dataWarehouse.ModelPatternDataDictionary[model][40],
-            //MaterialWheelsEnd = dataWarehouse.ModelPatternDataDictionary[model][41],
-            //SizeWheelsInitial = dataWarehouse.ModelPatternDataDictionary[model][42],
-            //SizeWheelsEnd = dataWarehouse.ModelPatternDataDictionary[model][43],
-            //TireInitial = dataWarehouse.ModelPatternDataDictionary[model][44],
-            //TireEnd = dataWarehouse.ModelPatternDataDictionary[model][45],
-
-            //ColorLightsInitial = dataWarehouse.ModelPatternDataDictionary[model][46],
-            //ColorLightsEnd = dataWarehouse.ModelPatternDataDictionary[model][47],
-            //PowerLightsInitial = dataWarehouse.ModelPatternDataDictionary[model][48],
-            //PowerLightsEnd = dataWarehouse.ModelPatternDataDictionary[model][49],
-
-            //PitchInitial = dataWarehouse.ModelPatternDataDictionary[model][50],
-            //PitchEnd = dataWarehouse.ModelPatternDataDictionary[model][51]
         };
 
         return pattern;
