+++ CarRental.PresentationLayer/Models/Car/Seeds/InteriorPattern.cs
@@ -4,18 +4,18 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CarRental.Data.Models.Car
+namespace CarRental.Data.Models.Car.Seeds
 {
-    public class TransmissionPattern
+    public class InteriorPattern
     {
         // PROPERTIES
 
         public int StatusInitialIndex { get; init; }
         public int StatusEndIndex { get; init; }
 
-        public int TypeInitial { get; init; }
-        public int TypeEnd { get; init; }
-        public int SpeedCountInitial { get; init; }
-        public int SpeedCountEnd { get; init; }
+        public int ColorInitial { get; init; }
+        public int ColorEnd { get; init; }
+        public int MaterialInitial { get; init; }
+        public int MaterialEnd { get; init; }
     }
 }
