+++ CarRental.PresentationLayer/Models/Car/Seeds/LightsPattern.cs
@@ -4,16 +4,18 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CarRental.Data.Models.Car
+namespace CarRental.Data.Models.Car.Seeds
 {
-    public class SignalPattern
+    public class LightsPattern
     {
         // PROPERTIES
 
         public int StatusInitialIndex { get; init; }
         public int StatusEndIndex { get; init; }
 
-        public int PitchInitial { get; init; }
-        public int PitchEnd { get; init; }
+        public int ColorInitial { get; init; }
+        public int ColorEnd { get; init; }
+        public int PowerInitial { get; init; }
+        public int PowerEnd { get; init; }
     }
 }
