+++ CarRental.PresentationLayer/Models/Car/Seeds/SignalPattern.cs
@@ -4,18 +4,16 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CarRental.Data.Models.Car
+namespace CarRental.Data.Models.Car.Seeds
 {
-    public class InteriorPattern
+    public class SignalPattern
     {
         // PROPERTIES
 
         public int StatusInitialIndex { get; init; }
         public int StatusEndIndex { get; init; }
 
-        public int ColorInitial { get; init; }
-        public int ColorEnd { get; init; }
-        public int MaterialInitial { get; init; }
-        public int MaterialEnd { get; init; }
+        public int PitchInitial { get; init; }
+        public int PitchEnd { get; init; }
     }
 }
